+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## master (unreleased)
 
 * Fix `ClassesBeforeIds` output format to handle `id` EnforcedStyle option better
+* Allow report_lint Rspec matcher to test lint message
 
 ## 0.18.2